feat: extract Layer protocol from BackgroundLayer
