+++ Sources/Clappr/Classes/Base/Layers/BackgroundLayer.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-final class BackgroundLayer: UIView {
+final class BackgroundLayer: UIView, Layer {
     init() {
         super.init(frame: .zero)
 
@@ -9,17 +9,5 @@ final class BackgroundLayer: UIView {
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-    
-    func attach(to view: UIView) {
-        view.addSubview(self)
-        
-        translatesAutoresizingMaskIntoConstraints = false
-        
-        widthAnchor.constraint(equalTo: view.widthAnchor).isActive = true
-        heightAnchor.constraint(equalTo: view.heightAnchor).isActive = true
-        
-        centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-        centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
     }
 }
