+++ Sources/Clappr/Classes/Base/Layers/Layer.swift
@@ -1,0 +1,19 @@
+import UIKit
+
+protocol Layer: UIView {
+    func attach(to view: UIView)
+}
+
+extension Layer {
+    func attach(to view: UIView) {
+        view.addSubview(self)
+        
+        translatesAutoresizingMaskIntoConstraints = false
+        
+        widthAnchor.constraint(equalTo: view.widthAnchor).isActive = true
+        heightAnchor.constraint(equalTo: view.heightAnchor).isActive = true
+        
+        centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
+        centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
+    }
+}
